fill BidListController and CurvePointController

+++ src/Domain/CurvePoint.cs
@@ -8,8 +8,8 @@ namespace Dot.Net.WebApi.Domain
         public int Id { get; set; }
         public int CurveId { get; set; }
         public DateTime AsOfDate { get; set; }
-        public double Term { get; set; }
-        public double Value { get; set; }
+        public decimal Term { get; set; }
+        public decimal Value { get; set; }
         public DateTime CreationDate { get; set; }
     }
 }
